app-api/requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.10
# by the following command:
#
#    pip-compile --output-file=app-api/requirements.txt app-api/requirements.in
#
aiofiles==24.1.0
    # via -r app-api/requirements.in
aiohappyeyeballs==2.4.6
    # via aiohttp
aiohttp==3.11.12
    # via -r app-api/requirements.in
aiosignal==1.3.2
    # via aiohttp
alembic==1.14.0
    # via -r app-api/requirements.in
annotated-types==0.7.0
    # via pydantic
anyio==4.7.0
    # via
    #   httpx
    #   openai
    #   starlette
async-property==0.2.2
    # via python-keycloak
async-timeout==5.0.1
    # via
    #   aiohttp
    #   redis
attrs==25.1.0
    # via aiohttp
cachetools==5.5.0
    # via -r app-api/requirements.in
certifi==2024.12.14
    # via
    #   httpcore
    #   httpx
    #   requests
cffi==1.17.1
    # via cryptography
charset-normalizer==3.4.0
    # via requests
click==8.1.7
    # via uvicorn
cryptography==44.0.1
    # via jwcrypto
deprecation==2.1.0
    # via python-keycloak
distro==1.9.0
    # via openai
exceptiongroup==1.2.2
    # via anyio
fastapi==0.115.6
    # via
    #   -r app-api/requirements.in
    #   fastapi-keycloak-middleware
fastapi-keycloak-middleware==1.1.0
    # via -r app-api/requirements.in
frozenlist==1.5.0
    # via
    #   aiohttp
    #   aiosignal
greenlet==3.1.1
    # via sqlalchemy
h11==0.14.0
    # via
    #   httpcore
    #   uvicorn
httpcore==1.0.7
    # via httpx
httpx==0.28.1
    # via
    #   openai
    #   python-keycloak
idna==3.10
    # via
    #   anyio
    #   httpx
    #   requests
    #   yarl
jiter==0.8.2
    # via openai
jwcrypto==1.5.6
    # via python-keycloak
mako==1.3.8
    # via alembic
markupsafe==3.0.2
    # via mako
multidict==6.1.0
    # via
    #   aiohttp
    #   yarl
openai==1.58.1
    # via -r app-api/requirements.in
packaging==24.2
    # via deprecation
pillow==11.0.0
    # via -r app-api/requirements.in
prometheus-client==0.21.1
    # via prometheus-fastapi-instrumentator
prometheus-fastapi-instrumentator==7.0.0
    # via -r app-api/requirements.in
propcache==0.3.0
    # via
    #   aiohttp
    #   yarl
psycopg2==2.9.10
    # via sqlalchemy
pycparser==2.22
    # via cffi
pydantic==2.10.4
    # via
    #   fastapi
    #   openai
pydantic-core==2.27.2
    # via pydantic
pyjwt==2.9.0
    # via redis
python-keycloak==4.7.3
    # via fastapi-keycloak-middleware
python-multipart==0.0.20
    # via -r app-api/requirements.in
pyyaml==6.0.2
    # via -r app-api/requirements.in
redis==5.3.0
    # via -r app-api/requirements.in
requests==2.32.3
    # via
    #   python-keycloak
    #   requests-toolbelt
requests-toolbelt==1.0.0
    # via python-keycloak
sniffio==1.3.1
    # via
    #   anyio
    #   openai
sqlalchemy[postgresql]==2.0.36
    # via
    #   -r app-api/requirements.in
    #   alembic
starlette==0.41.3
    # via
    #   fastapi
    #   prometheus-fastapi-instrumentator
tqdm==4.67.1
    # via openai
typing-extensions==4.12.2
    # via
    #   alembic
    #   anyio
    #   fastapi
    #   jwcrypto
    #   multidict
    #   openai
    #   pydantic
    #   pydantic-core
    #   sqlalchemy
    #   uvicorn
urllib3==2.2.3
    # via requests
uvicorn==0.34.0
    # via -r app-api/requirements.in
yarl==1.18.3
    # via aiohttp
